Move controller tests
