+++ pagespeed/kernel/sharedmem/BUILD
@@ -72,7 +72,7 @@ pagespeed_cc_test_library(
         "//test/pagespeed/kernel/base:kernel_test_util",
         "//test/pagespeed/kernel/cache:cache_test_base",
         "//test/pagespeed/kernel/thread:thread_test_base",
-        "//pagespeed/kernel/util:util_test_base",
+        "//test/pagespeed/kernel/util:util_test_base",
     ],
 )
