+++ test/pagespeed/controller/context_registry_test.cc
