+++ test/pagespeed/controller/expensive_operation_rpc_handler_test.cc
@@ -23,10 +23,10 @@
 
 #include "pagespeed/controller/controller.grpc.pb.h"
 #include "pagespeed/controller/controller.pb.h"
-#include "pagespeed/controller/grpc_server_test.h"
 #include "pagespeed/kernel/base/function.h"
 #include "pagespeed/kernel/base/string.h"
 #include "pagespeed/kernel/util/grpc.h"
+#include "test/pagespeed/controller/grpc_server_test.h"
 #include "test/pagespeed/kernel/base/gmock.h"
 #include "test/pagespeed/kernel/base/gtest.h"
 #include "test/pagespeed/kernel/thread/worker_test_base.h"
