+++ test/pagespeed/kernel/util/hashed_nonce_generator_test.cc
@@ -25,8 +25,8 @@
 #include "pagespeed/kernel/base/md5_hasher.h"
 #include "pagespeed/kernel/base/null_mutex.h"
 #include "pagespeed/kernel/util/nonce_generator.h"
-#include "pagespeed/kernel/util/nonce_generator_test_base.h"
 #include "test/pagespeed/kernel/base/gtest.h"
+#include "test/pagespeed/kernel/util/nonce_generator_test_base.h"
 
 namespace net_instaweb {
 namespace {
