+++ test/pagespeed/kernel/util/mock_nonce_generator.cc
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-#include "pagespeed/kernel/util/mock_nonce_generator.h"
+#include "test/pagespeed/kernel/util/mock_nonce_generator.h"
 
 #include "pagespeed/kernel/base/basictypes.h"
