+++ test/pagespeed/kernel/util/threadsafe_lock_manager_test.cc
@@ -37,12 +37,12 @@
 #include "pagespeed/kernel/base/thread_system.h"
 #include "pagespeed/kernel/base/timer.h"
 #include "pagespeed/kernel/thread/scheduler.h"
-#include "pagespeed/kernel/util/lock_manager_spammer.h"
 #include "pagespeed/kernel/util/platform.h"
 #include "test/pagespeed/kernel/base/gtest.h"
 #include "test/pagespeed/kernel/base/mock_timer.h"
 #include "test/pagespeed/kernel/base/named_lock_tester.h"
 #include "test/pagespeed/kernel/thread/mock_scheduler.h"
+#include "test/pagespeed/kernel/util/lock_manager_spammer.h"
 
 namespace net_instaweb {
